payment methods and dto shop service side

+++ shop1_service/src/main/java/top/shop/shop1_service/dto/PaymentDto.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
@@ -11,7 +12,7 @@ import java.time.LocalDateTime;
 @Setter
 public class PaymentDto {
    private String paymentUuid;
-   private double totalPrice;
+   private BigDecimal totalPrice;
    private LocalDateTime paymentDate;
    private int minutesBeforeExpiration;
 }
